Add sketchy way to do large read
Ultimately I should find a better way to do this, but I haven't yet
figured out how to achieve what I am looking for with a custom
script and weights in pgbench. I want to have either control over or
knowledge of exactly when the large read is starting and ending so I can
look at this in the output. I also would like to do a full transactional
workload simultaneously.